Move multer code
to ./controllers/api/imageRoutes.js
