+++ server.js
@@ -1,5 +1,4 @@
-// Import Node's built-in fs and path modules for working with files and directory or file paths
-const fs = require('fs');
+// Import Node's built-in path module for working with directory or file paths
 const path = require('path');
 
 // Import third-party package dependencies from the npm registry
@@ -7,12 +6,9 @@ require('dotenv').config();
 const express = require('express');
 const expressHandlebars = require('express-handlebars');
 const expressSession = require('express-session');
-const gm = require('gm');
-const multer = require('multer');
 const SequelizeStore = require('connect-session-sequelize')(
   expressSession.Store,
 );
-const uuid = require('uuid');
 
 const routes = require('./controllers');
 const sequelize = require('./config/connection');
@@ -40,31 +36,6 @@ const session = {
   }),
 };
 
-// Import lines 47-66 from https://www.digitalocean.com/community/tutorials/nodejs-uploading-files-multer-express
-// and https://expressjs.com/en/resources/middleware/multer.html with additional code comments and
-// changes for Hobby Tracker by Adam Hansen
-
-// Use the multer.diskStorage() method to tell Express where to store files to the disk
-const storage = multer.diskStorage({
-  // TODO: Define the destination folder where multer will temporarily store uploaded files
-  destination: function (req, file, callback) {
-    callback(null, '/src/my-images');
-  },
-  // TODO: Define the name of the uploaded file within the destination folder
-  // TODO: Use the uuid package to uniquely name the uploaded file
-  // TODO: Use the gm package and an installed GraphicsMagick instance to convert and resize the uploaded file
-  // TODO: Use Node's built-in fs and path modules to save the converted and resized file to a permanent destination folder
-  filename: function (req, file, callback) {
-    callback(null, file.fieldname);
-  },
-});
-
-const upload = multer({ storage: storage });
-
-// End code from https://www.digitalocean.com/community/tutorials/nodejs-uploading-files-multer-express
-// and https://expressjs.com/en/resources/middleware/multer.html with additional code comments and
-// changes for Hobby Tracker by Adam Hansen
-
 app.use(expressSession(session));
 
 app.engine('handlebars', handlebars.engine);
@@ -77,7 +48,5 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use(routes);
 
 sequelize.sync({ force: false }).then(() => {
-  app.listen(port, () => console.log('Now listening'));
+  app.listen(port, () => console.log(`Server is listening on port ${port}`));
 });
-
-
